api 단위테스트(router_unit) 커버리지 보강 중

+++ src/common/schemas/predict.py
@@ -8,5 +8,5 @@ class StockPredictionRequest(BaseModel):
 class StockPredictionResponse(BaseModel):
     symbol: str
     prediction: str
-    confidence: int
-    reason: str 
+    confidence: float # Changed from int to float
+    reason: str
